[WIP] (FM-7785) Additional test cleanup (#823)
move ipv6 tests into their own test files. one for happy path testing(checking parameters work), and one for exceptions or tests that check error output.
move missed ipv4 tests into appropriate test files.
